LPS-162959 Extract PopoverTooltip
